Bugfixes und Verbesserung der Bedienbarkeit

+++ WorkoutLog/src/com/workout/log/ManageTrainingDays.java
@@ -27,12 +27,14 @@ import com.workout.log.bo.Exercise;
 import com.workout.log.bo.PerformanceTarget;
 import com.workout.log.bo.TrainingDay;
 import com.workout.log.bo.Workoutplan;
+import com.workout.log.data.Default;
 import com.workout.log.db.PerformanceTargetMapper;
 import com.workout.log.db.TrainingDayMapper;
 import com.workout.log.db.WorkoutplanMapper;
 import com.workout.log.dialog.TrainingDayAddDialogFragment;
 import com.workout.log.dialog.TrainingDayUpdateDialogFragment;
 import com.workout.log.fragment.TrainingDaysSearchBarFragment;
+import com.workout.log.listAdapter.DefaultAddListAdapter;
 import com.workout.log.listAdapter.TrainingDayListAdapter;
 
 public class ManageTrainingDays extends Fragment implements OnItemClickListener, OnItemLongClickListener, UndoBarController.UndoListener{
@@ -91,7 +93,7 @@ public class ManageTrainingDays extends Fragment implements OnItemClickListener,
 		trainingDayListView.setOnItemClickListener(this);
 		trainingDayListView.setOnItemLongClickListener(this);
 		
-		updateListView(null);
+		updateListView(tdMapper.getAllTrainingDay(), null);
 		loadSwipeToDismiss();  
 		setHasOptionsMenu(true);
 	}
@@ -106,13 +108,18 @@ public class ManageTrainingDays extends Fragment implements OnItemClickListener,
 	public boolean onOptionsItemSelected(MenuItem item) {
 		int itemId = item.getItemId();
 		if (itemId == R.id.menu_add) {
-			TrainingDayAddDialogFragment dialogFragment = TrainingDayAddDialogFragment.newInstance(getActivity(), trainingDayListAdapter);
-			dialogFragment.show(this.getFragmentManager(), "Open Exercise Add Dialog on Click");
+			showDialogAddFragment(null);
 		}
 
 		return super.onOptionsItemSelected(item);
 	}
-
+	
+	public void showDialogAddFragment(String trainingDayStringName){
+		TrainingDayAddDialogFragment dialogFragment = TrainingDayAddDialogFragment.newInstance(getActivity(),
+				trainingDayListAdapter, trainingDayStringName, this);
+		dialogFragment.show(this.getFragmentManager(), "Open Exercise Add Dialog on Click");
+	}
+	
 	@Override
 	public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
 		TrainingDay td = (TrainingDay) parent.getItemAtPosition(position);
@@ -154,8 +161,30 @@ public class ManageTrainingDays extends Fragment implements OnItemClickListener,
 	}
 
 	@SuppressWarnings("unchecked")
-	public void updateListView(ArrayList<TrainingDay> trainingDayList) {
-		new BackGroundTask().execute(trainingDayList);
+	public void updateListView(ArrayList<TrainingDay> trainingDayList, final String trainingDayStringName) {
+		if (trainingDayList.size() == 0){
+			Default d = new Default();
+			if (trainingDayStringName == null || trainingDayStringName.equals("")){
+				d.setTitel(getString(R.string.noTrainingDay));
+			}else{
+				d.setTitel(trainingDayStringName + " " + getResources().getString(R.string.NotFound));
+				d.setHint(getResources().getString(R.string.Add));
+			}
+			ArrayList<Default> ld = new ArrayList<Default>();
+			ld.add(d);
+			DefaultAddListAdapter l = new DefaultAddListAdapter(getActivity(), 0, ld);
+			trainingDayListView.setAdapter(l);
+			trainingDayListView.setOnItemClickListener(new OnItemClickListener(){
+				@Override
+				public void onItemClick(AdapterView<?> arg0, View arg1,int arg2, long arg3) {
+					showDialogAddFragment(trainingDayStringName);
+				}	
+			});
+			
+		}else{
+			trainingDayListView.setOnItemClickListener(this);
+			new BackGroundTask().execute(trainingDayList);
+		}
 	}
 	
 	/**
@@ -193,8 +222,10 @@ public class ManageTrainingDays extends Fragment implements OnItemClickListener,
                 		PerformanceTargetMapper pMapper = new PerformanceTargetMapper(getActivity());
                 		
                 		tMapper.delete(trainingDay);
-                		for (Exercise e : trainingDay.getExerciseList()){
-                			tMapper.deleteExerciseFromTrainingDay(trainingDay.getId(), e);
+                		if (trainingDay.getExerciseList() != null){
+	                		for (Exercise e : trainingDay.getExerciseList()){
+	                			tMapper.deleteExerciseFromTrainingDay(trainingDay.getId(), e);
+	                		}
                 		}
                 		tMapper.deleteTrainingDayFromAllWorkoutplan(trainingDay.getId());
                 		pMapper.deleteTrainingDayFromAllPerformanceTarget(trainingDay.getId());
@@ -315,7 +346,9 @@ public class ManageTrainingDays extends Fragment implements OnItemClickListener,
 	    protected void onPostExecute(TrainingDayListAdapter result) {
 	        super.onPostExecute(result);
 	        
-	        if (result != null) trainingDayListView.setAdapter(result);
+	        if (result != null){
+	        	trainingDayListView.setAdapter(result);
+	        }
 	        
 	        getActivity().setProgressBarIndeterminateVisibility(false);
 	    }
